+++ packages/gatsby-cli/src/init-starter.ts
@@ -11,9 +11,8 @@ import prompts from "prompts"
 import url from "url"
 import { updateSiteMetadata } from "gatsby-core-utils"
 import report from "./reporter"
-import { getPackageManager, promptPackageManager } from "./util/package-manager"
-import { isTTY } from "./util/is-tty"
-import reporter from "../lib/reporter"
+import { getPackageManager, setPackageManager } from "./util/package-manager"
+import reporter from "./reporter"
 
 const spawnWithArgs = (
   file: string,
@@ -29,26 +28,13 @@ const spawn = (
   const [file, ...args] = cmd.split(/\s+/)
   return spawnWithArgs(file, args, options)
 }
-// Checks the existence of yarn package and user preference if it exists
+// Checks the existence of yarn package
 // We use yarnpkg instead of yarn to avoid conflict with Hadoop yarn
 // Refer to https://github.com/yarnpkg/yarn/issues/673
-const shouldUseYarn = async (): Promise<boolean> => {
+const checkForYarn = async (): Promise<boolean> => {
   try {
     execSync(`yarnpkg --version`, { stdio: `ignore` })
-
-    let packageManager = getPackageManager()
-    if (!packageManager) {
-      // if package manager is not set:
-      //  - prompt user to pick package manager if in interactive console
-      //  - default to yarn if not in interactive console
-      if (isTTY()) {
-        packageManager = (await promptPackageManager()) || `yarn`
-      } else {
-        packageManager = `yarn`
-      }
-    }
-
-    return packageManager === `yarn`
+    return true
   } catch (e) {
     return false
   }
@@ -114,9 +100,23 @@ const install = async (rootPath: string): Promise<void> => {
   report.info(`Installing packages...`)
   process.chdir(rootPath)
 
+  const npmConfigUserAgent = process.env.npm_config_user_agent
+
   try {
-    if (await shouldUseYarn()) {
-      await fs.remove(`package-lock.json`)
+    if (!getPackageManager()) {
+      if (npmConfigUserAgent?.includes(`yarn`)) {
+        setPackageManager(`yarn`)
+      } else {
+        setPackageManager(`npm`)
+      }
+    }
+    if (getPackageManager() === `yarn` && checkForYarn()) {
+      if (await fs.pathExists(`package-lock.json`)) {
+        if (!(await fs.pathExists(`yarn.lock`))) {
+          await spawn(`yarnpkg import`)
+        }
+        await fs.remove(`package-lock.json`)
+      }
       await spawn(`yarnpkg`)
     } else {
       await fs.remove(`yarn.lock`)