+++ packages/gatsby-admin/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.154](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.153...gatsby-admin@0.1.154) (2020-09-24)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.153](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.152...gatsby-admin@0.1.153) (2020-09-23)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.152](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.151...gatsby-admin@0.1.152) (2020-09-18)
+
+**Note:** Version bump only for package gatsby-admin
+
+## [0.1.151](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.150...gatsby-admin@0.1.151) (2020-09-16)
+
+### Features
+
+- **gatsby-admin:** prompt to restart develop process on config changes ([#26916](https://github.com/gatsbyjs/gatsby/issues/26916)) ([1d3ff02](https://github.com/gatsbyjs/gatsby/commit/1d3ff024cb349906373657e8e826a2ac6a415c7c))
+
+## [0.1.150](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.149...gatsby-admin@0.1.150) (2020-09-16)
+
+### Features
+
+- **gatsby-admin:** track errors ([#26903](https://github.com/gatsbyjs/gatsby/issues/26903)) ([617cc1d](https://github.com/gatsbyjs/gatsby/commit/617cc1dd0e16d60abbea1caec8d3a63f8dbe4142))
+- **gatsby-admin:** track pageviews to telemetry ([#26902](https://github.com/gatsbyjs/gatsby/issues/26902)) ([f61b19b](https://github.com/gatsbyjs/gatsby/commit/f61b19be4779f37e959d5169c40299bbd655585a))
+
 ## [0.1.149](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.1.148...gatsby-admin@0.1.149) (2020-09-15)
 
 **Note:** Version bump only for package gatsby-admin